+++ test/karma/mock-fetch-results.js
@@ -12,9 +12,16 @@ let originalFetch = window.fetch;
 function mockFetchResults(results) {
   window.fetch =
     (url, ...otherParams) => new Promise((resolve, _) => {
-      const mockedResult = results?.find(r => r[0].test(url))?.[1]
-      if(mockedResult) {
-        resolve({ json: () => mockedResult });
+      const mockedItem = results?.find(
+        r => r[0] instanceof RegExp ? r[0].test(url) : url.indexOf(r[0]) !== -1
+      );
+      const okResult = mockedItem?.[1];
+      const badResult = mockedItem?.[2];
+
+      if(okResult) {
+        resolve({ json: () => okResult, ok: true });
+      } else if(badResult) {
+        resolve({ json: () => badResult, ok: false });
       } else {
         resolve(originalFetch.call(window, url, ...otherParams)
           .then(r => r.json())
@@ -22,7 +29,7 @@ function mockFetchResults(results) {
             console.error(`"${url}" is not mocked and returns:\n${JSON.stringify(bundle, null, 2)}`);
             return {
               json: () => bundle
-            }
+            };
           }));
       }
     });